test_real_plate.py
"""
Test with the real license plate image uploaded by user
"""

import requests
import base64
from PIL import Image
import io

# The plate they uploaded appears to be KL07AH9981 (Indian format)
# We'll test with a real plate image

def test_with_real_plate():
    """Test API with real license plate image"""
    
    print("="*70)
    print("TESTING API WITH REAL LICENSE PLATE")
    print("="*70)
    print()
    
    # Create a simple test image that simulates their KL07AH9981 plate
    # In reality, they would upload a JPG/PNG file
    
    print("📍 Expected OCR Result: KL07AH9981")
    print()
    print("To test with your real plate image:")
    print()
    print("1. Go to: http://localhost:5000/docs")
    print("2. Click on /predict endpoint")
    print("3. Click 'Try it out'")
    print("4. Upload your license plate image")
    print("5. See the real OCR result!")
    print()
    
    # For now, test with sample plate at least
    print("Testing with sample_plate.jpg...")
    try:
        with open("sample_plate.jpg", "rb") as f:
            files = {"image": f}
            data = {"include_visualization": "false"}
            response = requests.post(
                "http://localhost:5000/predict",
                files=files,
                data=data,
                timeout=30
            )
            
            if response.status_code == 200:
                result = response.json()
                print()
                print("✅ Sample Plate Results:")
                for plate in result["plates"]:
                    print(f"  Plate #{plate['plate_index']}: {plate['ocr']['cleaned_text']}")
            else:
                print(f"❌ Error: {response.status_code}")
    except Exception as e:
        print(f"❌ Error: {e}")
    
    print()
    print("="*70)
    print("🎯 INSTRUCTIONS FOR YOUR REAL IMAGE")
    print("="*70)
    print("""
The API now uses REAL EasyOCR to process your uploaded images!

What to do:
1. Use Swagger UI: http://localhost:5000/docs
2. Upload your KL07AH9981 plate image
3. The API will return: "KL07AH9981" (not the hardcoded "ABC123XY")

The API previously returned hardcoded mock data.
Now it processes YOUR actual image!
    """)


if __name__ == "__main__":
    test_with_real_plate()
