requirements.txt
# Core Computer Vision & ML
numpy==1.24.3
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
tensorflow==2.13.0
tensorflow-hub==0.14.0
scipy==1.11.3
scikit-image==0.21.0

# OCR
easyocr==1.7.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
tqdm==4.66.1
pandas==2.0.3
matplotlib==3.8.0

# Development (optional)
ipython==8.12.0
jupyter==1.0.0
